day 16 - vs compiler switches

+++ source/handmade.cpp
@@ -23,8 +23,8 @@ RenderWeirdGradient(game_offscreen_buffer* Buffer, int BlueOffset, int GreenOffs
 			X < Buffer->Width;
 			++X)
 		{
-			uint8 Blue = (X + BlueOffset);
-			uint8 Green = (Y + GreenOffset);
+			uint8 Blue = (uint8)(X + BlueOffset);
+			uint8 Green = (uint8)(Y + GreenOffset);
 
 			*Pixel++ = ((Green << 8) | Blue);
 		}
@@ -85,8 +85,8 @@ internal void GameUpdateAndRender(game_memory* Memory, game_input* Input, game_o
 	if (Input0->IsAnalog)
 	{
 		// NOTE(casey): Use analog movement tuning
-		GameState->BlueOffset += (int)4.0f * (Input0->EndX);
-		GameState->ToneHz = 256 + (int)(128.0f * (Input0->EndY));
+		GameState->BlueOffset += (int)(4.0f * Input0->EndX);
+		GameState->ToneHz = 256 + (int)(128.0f * Input0->EndY);
 	}
 	else
 	{
